devtools: add go-dirhash-native
When we are building to applciations, we sometimes need to patch
or otherise make changes that will change the checksum. If a module
cache or other validated/checked method is being used, you need an
updated checksum.

Using this utility will ensure that the same hash / checksum routine
as the go build will use .. is used for an updated sum.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>