go-mod-discovery: move patching after discovery build
When updating a SRCREV and then running discovery, we expect that
some patches will fail. We don't want that to block discovery as
these patches are not normally ADDING dependencies that would be
found.

So we move discovery before patch and patch failures can be dealt
with on a full build.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>